--- frontend/app/src/components/wizard/types.ts
@@ -1,7 +1,0 @@
-import type { Readable } from "svelte/store";
-
-export type WizardState = {
-    currentStep: Readable<number>;
-    next: () => void;
-    previous: () => void;
-};
